oci: unpack: slurp up raw layer stream before Close()
It's possible there's some trailing bytes after the gzip stream (which
truth be told, probably shouldn't be there), so in order to avoid
spurious errors on layer unpacking with slightly-broken images, just
slurp up that data.

But we should output logging information since this indicates that the
original image builder probably has a bug of some kind.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>